feat: OpenFold3 defaults to the partial bfloat16 track
`dtype` defaults to `bfloat16` and `--option dtype=float32` restores upstream's
`32-true` inference precision. `confidence_dtype` follows `dtype`, so one word
moves both regions.

A 28-row panel over three targets -- one input, one checkpoint, 10 recycles,
200 diffusion steps, five samples per seed, from one source tree -- measured
bfloat16 faster and smaller at every size: 98.7 -> 76.2 s and 9,273 -> 5,552
MiB at 1,003 tokens, 402-405 -> 262-263 s and 25,066 -> 19,107 MiB at 2,096,
954 -> 661 s and 50,412 -> 34,893 MiB at 3,012. Per-chain RMSD against the
deposited coordinates is the same on both arms at all three sizes.

**The 3,012-token result that kept this opt-in was read wrongly, twice.** It
was measured as a same-index residual between the bfloat16 arm and a float32
arm at the same seed -- 25 A at seed 202 -- and the float32 arm's healthy
within-set spread (0.569 A) was read as "the control is fine, so it is the
arm". A residual between two arms cannot say which of them moved, and a
within-arm spread cannot see a basin the whole arm is sitting in: five samples
agree with each other just as tightly in the wrong place as in the right one.

Scored against the deposited structure instead, on a permutation-aware chain
assignment, with the catalase core (122-753) fitted and the N-terminal arm
(27-121) read separately, over twelve 3,012-token arms of five samples: the
core is 0.55-1.03 A in all twelve, and the arm misses its basin in exactly two
-- float32 at seed 202 (56.41-56.66 A) and bfloat16 at seed 404 (56.67-56.79
A), each in all five samples of that seed. Each dtype misses at one seed in
six, and neither misses at a seed the other one makes. Six seeds do not
estimate a rate; what they rule out is that the miss belongs to `bfloat16`.
The mechanism is not measured -- it is consistent with this port's per-recycle
MSA resubsampling (`backends/openfold3.py:636`).

The criterion for an execution default here is whether the structure holds,
not whether the arithmetic matches upstream's. Parity is the exception, so
five sites pin `dtype="float32"`: `tests/parity/test_openfold3.py`,
`bench/openbind_core_replay.py` which it copies, and the three
`tests/models/openfold3` suites that compare against upstream's float32
forward on an uncast float32 state dict. `bench/spec.py` pins it too, so the
FoldJAX-versus-upstream matrix keeps comparing two implementations at one
precision rather than two precisions.

`cast_narrow_params` was an identity under the old default and is
load-bearing now: every entry point that loads a checkpoint has to call it, or
activations round down at the region boundary and promote back at the first
matmul -- the entry rounding paid and nothing running narrow.

Every OpenFold3 compilation-cache namespace moves once, and that is the flip's
own consequence rather than a separate change. `cache_profile` now records
`dtype` and `confidence_dtype` as the widths the run resolves to, always,
instead of stripping a value equal to the default. Stripping would make
*absence* the signature for the shipped bfloat16 -- and absence already means
float32, which is what every run recorded before today wrote. Two programs,
one name. The aliasing this replaces still holds: an omitted `dtype` and an
explicit `bfloat16` name one namespace, and only `float32` forks. Boltz-2's
`pair_residual_dtype` took the same shape yesterday for the same reason. The
cost is one cold compile per shape.

Boltz-2's `matmul_precision` gets the same treatment in the same place, and it
is the only other port that needs it: it is the one whose pinned value changed
(`highest` until 2026-09-11, and the knob was not in `compile_options` at all
then), so absence in a recorded run means `highest` and cannot be reused for
the new `high`. OpenFold3 and Protenix pin `high` and always have, so their
strip is safe. The census merge left a duplicate `matmul_precision` key in
Boltz-2's released-defaults table; it is gone.

`models/openfold3/dtype.py`'s account is rewritten rather than annotated. The
seed-dependence table, the "it is the arm, not the harness" reading and the
"why 3,012 and not 2,096" search are deleted: the question is answered and the
answer is that it is not the dtype. `docs/openfold3-bf16-default-evidence-2026-09-12.md`
carries the panel and its limits.

+++ bench/openbind_core_replay.py
@@ -248,10 +248,16 @@ def main(argv=None):
         features = streamed_host_features(features)
         host_steps = HOST_FEATURE_STEPS
     budget = array_budget_bytes(streamed=args.streamed, all_arrays=args.all_arrays)
+    # `dtype="float32"` because every capture this replays is a float32
+    # native run and `map_inference_params` above hands back float32
+    # parameters uncast: under the port's partial-bfloat16 default the
+    # activations would narrow, meet those parameters and promote back, so
+    # the replay would pay an entry rounding and run nothing narrow.
+    # tests/parity/test_openfold3.py copies this call and pins the same value.
     config = released_config(
         n_token=features["token_mask"].shape[-1],
         n_atom=features["atom_mask"].shape[-1],
-        num_recycles=4, num_samples=5, num_steps=200,
+        num_recycles=4, num_samples=5, num_steps=200, dtype="float32",
         msa_depth=tape.msa_indices.shape[1], max_array_bytes=budget,
         returned_representations=(
             ("single_inputs", "single", "pair") if args.capture_trunk else ()
